update number of active filters
