+++ src/common/index.ts
@@ -1,11 +1,15 @@
 export * from "./src/Events";
-export * from "./src/Keywords";
-export * from "./src/MessageCommands";
-export * from "./src/PostMessageSchema";
-export * from "./src/MessageSchemaMap";
+export * from "./src/model/Keywords";
+export * from "./src/postMessage/MessageCommands";
+export * from "./src/postMessage/PostMessageSchema";
+export * from "./src/postMessage/MessageSchemaMap";
+export * from "./src/postMessage/CommandId";
 export * from "./src/model";
 export * from "./src/service/PostMessageServiceBase";
 export type { IPostMessageService } from "./src/service/IPostMessageService";
+
+
+
