+++ src/common/src/Events.ts
@@ -2,38 +2,6 @@ import { z } from "zod";
 export const LogLevelSchema = z.union([z.literal("info"), z.literal("debug"), z.literal("warning"), z.literal("error")]);
 export type LogLevel = z.TypeOf<typeof LogLevelSchema>;
 
-export type DisplaySettingsChangedEvent = {
-    /**
-     * Whether to display the file names. A null value means that the user did not change the setting.
-     */
-    displayFileNames: boolean | null;
-
-    /**
-     * Whether to display the dates in line. A null value means that the user did not change the setting.
-     */
-    displayDatesInLine: boolean | null;
-
-    /**
-     * Whether to display the guids. A null value means that the user did not change the setting.
-     */
-    displayGuids: boolean | null;
-};
-
-export type FilterKeywordChangedEvent = {
-    /**
-     * The id of the checkbox.
-     */
-    checkboxId: string;
-    /**
-     * The keyword filter.
-     */
-    keyword: string;
-    /**
-     * The state of the checkbox.
-     */
-    isChecked: boolean;
-};
-
 /**
  * The event that is fired when the filter changes.
  */
