+++ src/common/src/model/index.ts
@@ -1,1 +1,3 @@
 export * from "./SummaryInfo";
+
+export * from "./Keywords";
