+++ src/common/src/postMessage/PostMessageSchema.ts
@@ -1,25 +1,5 @@
 import { z } from "zod";
-
-/**
- * The valid commands that can be sent between extension and webview.
- */
-export const CommandIdSchema = z.union([
-    z.literal("logMessage"),
-    z.literal("logErrorMessage"),
-    z.literal("filterCheckboxStateChange"),
-    z.literal("filterLogLevel"),
-    z.literal("filterTime"),
-    z.literal("filterSessionId"),
-    z.literal("filterNoEventTime"),
-    z.literal("updateNumberOfActiveFilters"),
-    z.literal("getSummary"),
-    z.literal("getSummaryResponse"),
-    z.literal("displaySettingsChanged"),
-    z.literal("openLogsDocument"),
-    z.literal("keywordHighlightStateChange"),
-    z.literal("updateTimeFilters"),
-    z.literal("messageAck"),
-]);
+import { CommandIdSchema } from "./CommandId";
 
 /**
  * The valid commands that can be sent between extension and webview.
@@ -50,7 +30,4 @@ export const PostMessageSchema = z.object({
  * The schema for the message that is sent between the webview and the extension with unknown data.
  */
 export type PostMessageWithUnknownData = z.TypeOf<typeof PostMessageSchema>;
-
-
-
